requirements.txt
pytest==7.4.0
playwright~=1.51.0
pytest-playwright==0.4.0
pytest-html==3.2.0
python-dotenv==1.0.0
pytest-xdist==3.3.1
